+++ egov-filestore/src/main/java/org/egov/filestore/validator/StorageValidator.java
@@ -11,6 +11,7 @@ import org.egov.filestore.domain.model.Artifact;
 import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
 
 @Component
 public class StorageValidator {
@@ -27,9 +28,10 @@ public class StorageValidator {
 
 	public void validate(Artifact artifact) {
 			
-		String extension = FilenameUtils.getExtension(artifact.getMultipartFile().getOriginalFilename());
+		String extension = (FilenameUtils.getExtension(artifact.getMultipartFile().getOriginalFilename())).toLowerCase();
 		validateFileExtention(extension);
 		validateContentType(artifact.getFileContentInString(), extension);
+		validateInputContentType(artifact);
 	}
 	
 	private void validateFileExtention(String extension) {
@@ -56,6 +58,19 @@ public class StorageValidator {
 							+ fileStoreConfig.getAllowedKeySet());
 		}
 	}
+
+	private void validateInputContentType(Artifact artifact){
+
+		MultipartFile file =  artifact.getMultipartFile();
+		String contentType = file.getContentType();
+		String extension = (FilenameUtils.getExtension(artifact.getMultipartFile().getOriginalFilename())).toLowerCase();
+
+
+		if (!fileStoreConfig.getAllowedFormatsMap().get(extension).contains(contentType)) {
+			throw new CustomException("EG_FILESTORE_INVALID_INPUT", "Invalid Content Type");
+		}
+	}
+
 	
 	/*private void validateFilesToUpload(List<MultipartFile> filesToStore, String module, String tag, String tenantId) {
 		if (CollectionUtils.isEmpty(filesToStore)) {