+++ egov-idgen/src/main/java/org/egov/id/model/IdRequest.java
@@ -2,6 +2,7 @@ package org.egov.id.model;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -20,14 +21,17 @@ import lombok.*;
 @NoArgsConstructor
 public class IdRequest {
 
+	@Size(max = 200)
 	@JsonProperty("idName")
 	@NotNull
 	private String idName;
 
 	@NotNull
+	@Size(max=200)
 	@JsonProperty("tenantId")
 	private String tenantId;
 
+	@Size(max = 200)
 	@JsonProperty("format")
 	private String format;