+++ egov-searcher/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.1.2 - 26-02-2021
+
+- Minor bug fixes for new operators
+
+## 1.1.1 - 13-01-2021
+
+- added Touppercase, LIKE, ILIKE and TOLOWERCASE as operators
+
+## 1.1.1
+
 ## 1.1.0
 
 - Add support for Bill Genie search APIs